movendo e documentnado AMQPRemoteException
